requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
SQLAlchemy==2.0.32
PyMySQL==1.1.1
python-dotenv==1.0.1
click==8.1.7